attempt to add installer
