+++ Makefile
@@ -1,12 +1,13 @@
 export ROOT_DIR := $(shell dirname $(realpath $(lastword $(MAKEFILE_LIST))))
-BUILD_DIR = $(ROOT_DIR)/build
+SD_DIR = $(ROOT_DIR)/stringdecomposer
+BUILD_DIR = $(SD_DIR)/build
 BIN_DIR = $(BUILD_DIR)/bin
-SRC_DIR = $(ROOT_DIR)/src
+SRC_DIR = $(SD_DIR)/src
 
-TEST_QUERY = "test_data/read.fa"
-TEST_MONOMERS = "test_data/DXZ1_star_monomers.fa"
-TEST_OUTDIR = "test_data"
-TEST_REFERENCE = "test_data/final_decomposition_fc89af8.tsv"
+TEST_QUERY = $(SD_DIR)/test_data/read.fa
+TEST_MONOMERS = $(SD_DIR)/test_data/DXZ1_star_monomers.fa
+TEST_OUTDIR = $(SD_DIR)/test_data
+TEST_REFERENCE = $(SD_DIR)/test_data/final_decomposition_fc89af8.tsv
 
 build:
 	mkdir -p $(BIN_DIR)
@@ -28,7 +29,10 @@ test_launch_install: install
 
 clean:
 	-rm -rf $(BUILD_DIR)
-	-rm -rf test_data/final_decomposition{_alt.tsv,_raw.tsv,.tsv} test_data/stringdecomposer.log
+	-rm -rf $(SD_DIR)/test_data/final_decomposition_alt.tsv
+	-rm -rf $(SD_DIR)/test_data/final_decomposition_raw.tsv
+	-rm -rf $(SD_DIR)/test_data/final_decomposition.tsv
+	-rm -rf $(SD_DIR)/test_data/stringdecomposer.log
 	@if [ -f install_footprint.txt ]; then\
 		echo "removing install footprint from install_footprint.txt";\
 		cat install_footprint.txt | xargs rm -rf;\
