+++ stringdecomposer/__version__.py
