+++ stringdecomposer/git.py
