+++ stringdecomposer/src/main.cpp
