+++ stringdecomposer/standard_logger.py
