+++ stringdecomposer/test_data/DXZ1_star_monomers.fa
