+++ stringdecomposer/test_data/read.fa
